Make the package ClosureTable and Entity abstract
The classes that come with the package shouldn't force extra things on
the user like timestamps and soft deletes, it should work with or
without them.

The tests have been updated so that they don't create instances of the
classes from the package but instead the tests classes deal with that.